server/pyproject.toml
[tool.poetry]
name = "marron_machine"
version = "0.0.1"
description = "Marron-machine"
authors = ["Maruo.S <maruo@marusoftware.net>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10.0"
fastapi = "^0.110.0"
tortoise-orm = "^0.21.7"
fastapi-sso = "^0.12.1"
pydantic = {extras = ["email"], version = "^2.6.2"}
pydantic-settings = "^2.2.1"
itsdangerous = "^2.1.2"
aerich = "^0.7.2"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
ansible-runner = "^2.4.0"
aionetbox = "^1.7.4"
passlib = "^1.7.4"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
fastapi = {extras = ["dev"], version = "^0.110.0"}
watchfiles = "^0.21.0"
tortoise-cli = "^0.1.2"


[tool.poetry.group.postgres.dependencies]
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.7"}



[tool.poetry.group.test.dependencies]
pytest = "^8.3.3"

[tool.aerich]
tortoise_orm = "marron_machine.db.DB_CONFIG"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
